Notes screen and switch codebeat v1,
